move big code from api.js to tools
