+++ api.js
@@ -48,8 +48,6 @@ const certsRoutes = require('./lib/api/certs');
 const webhooksRoutes = require('./lib/api/webhooks');
 const settingsRoutes = require('./lib/api/settings');
 const { SettingsHandler } = require('./lib/settings-handler');
-const fs = require('fs');
-const yaml = require('js-yaml');
 
 let userHandler;
 let mailboxHandler;
@@ -580,350 +578,11 @@ module.exports = done => {
         tools.responseWrapper(async (req, res) => {
             res.charSet('utf-8');
 
-            let docs = `
-openapi: 3.0.0
-info:
-    title: WildDuck API
-    description: WildDuck API docs
-    version: 1.0.0
-    contact:
-        url: 'https://github.com/nodemailer/wildduck'
-
-servers:
-    - url: 'https://api.wildduck.email'
-
-tags:
-    - name: Addresses
-    - name: ApplicationPasswords
-    - name: Archive
-      description: Archive includes all deleted messages. Once messages are old enough then these are permanenetly deleted from the archive as well. Until then you can restore the deleted messages.
-    - name: Audit
-      description: 'Auditing allows to monitor an email account. All existing, deleted and new emails are copied to the auditing system. See also https://github.com/nodemailer/wildduck-audit-manager'
-    - name: Authentication
-    - name: Autoreplies
-    - name: Certs
-      description: WildDuck allows to register TLS certificates to be used with SNI connections. These certificates are used by IMAP, POP3, API and SMTP servers when a SNI capable client establishes a TLS connection. This does not apply for MX servers.
-    - name: DKIM
-      description: Whenever an email is sent WildDuck checks if there is a DKIM key registered for the domain name of the sender address and uses it to sign the message.
-    - name: DomainAccess
-      description: Add sender domain names to allowlist (messages are all accepted) or blocklist (messages are sent to Spam folder)
-    - name: DomainAliases
-    - name: Filters
-    - name: Mailboxes
-    - name: Messages
-    - name: Settings
-    - name: Storage
-      description: Storage allows easier attachment handling when composing Draft messages. Instead of uploading the attachmnent with every draft update, you store the attachment to the Storage and then link stored file for the Draft.
-    - name: Submission
-    - name: TwoFactorAuth
-    - name: Users
-    - name: Webhooks\n`;
-            const mapPathToMethods = {}; // map -> {path -> {post -> {}, put -> {}, delete -> {}, get -> {}}}
-
             const routes = server.router.getRoutes();
-            for (const routePath in routes) {
-                const route = routes[routePath];
-                const { spec } = route;
 
-                if (!spec.include) {
-                    continue;
-                }
-
-                if (!mapPathToMethods[spec.path]) {
-                    mapPathToMethods[spec.path] = {};
-                }
-
-                mapPathToMethods[spec.path][spec.method.toLowerCase()] = {};
-                const methodObj = mapPathToMethods[spec.path][spec.method.toLowerCase()];
-                // 1) add tags
-                methodObj.tags = spec.tags;
-
-                // 2) add summary
-                methodObj.summary = spec.summary || '';
-
-                // 3) add description
-                methodObj.description = spec.description || '';
-
-                // 4) add operationId
-                methodObj.operationId = spec.name || route.name;
-
-                // 5) add requestBody
-                const applicationType = spec.applicationType || 'application/json';
-                methodObj.requestBody = {
-                    content: {
-                        [applicationType]: {
-                            schema: {
-                                type: 'object',
-                                properties: {}
-                            }
-                        }
-                    },
-                    required: true
-                };
-
-                for (const reqBodyKey in spec.requestBody) {
-                    const reqBodyKeyData = spec.requestBody[reqBodyKey];
-
-                    parseJoiObject(reqBodyKey, reqBodyKeyData, methodObj.requestBody.content[applicationType].schema.properties);
-                }
-
-                // 6) add parameters (queryParams + pathParams).
-                methodObj.parameters = {};
-                for (const paramKey in spec.pathParams) {
-                    const paramKeyData = spec.pathParams[paramKey];
-
-                    methodObj.parameters[paramKey] = {};
-                    const obj = methodObj.parameters[paramKey];
-                    obj.in = 'path';
-                    obj.description = paramKeyData._flags.description || '';
-                    obj.required = paramKeyData._flags.presence === 'required';
-                    obj.schema = { type: paramKeyData.type };
-                }
-
-                for (const paramKey in spec.queryParams) {
-                    const paramKeyData = spec.queryParams[paramKey];
-
-                    methodObj.parameters[paramKey] = {};
-                    const obj = methodObj.parameters[paramKey];
-                    obj.in = 'query';
-                    obj.description = paramKeyData._flags.description || '';
-                    obj.required = paramKeyData._flags.presence === 'required';
-                    obj.schema = { type: paramKeyData.type };
-                }
-
-                // 7) add responses
-                methodObj.responses = {};
-
-                for (const resHttpCode in spec.response) {
-                    const restBodyData = spec.response[resHttpCode];
-
-                    parseJoiObject(resHttpCode, restBodyData, methodObj.responses);
-                }
-            }
-
-            const components = { components: { schemas: {} } };
-
-            for (const path in mapPathToMethods) {
-                // for every path
-                const pathData = mapPathToMethods[path];
-
-                for (const httpMethod in pathData) {
-                    // for every http method (post, put, get, delete)
-                    const innerData = pathData[httpMethod];
-
-                    // for every requestBody obj
-                    for (const key in innerData.requestBody.content[Object.keys(innerData.requestBody.content)[0]].schema.properties) {
-                        const reqBodyData = innerData.requestBody.content[Object.keys(innerData.requestBody.content)[0]].schema.properties[key];
-
-                        parseComponetsDecoupled(reqBodyData, components.components.schemas);
-                        replaceWithRefs(reqBodyData);
-                    }
-
-                    // for every response object
-                    for (const key in innerData.responses) {
-                        // key here is http method (2xx, 4xx, 5xx)
-                        const obj = innerData.responses[key];
-                        parseComponetsDecoupled(obj, components.components.schemas);
-                        replaceWithRefs(obj);
-                    }
-                }
-            }
-
-            // refify components that use other components
-            for (const obj of Object.values(components.components.schemas)) {
-                replaceWithRefs(obj);
-            }
-
-            const finalObj = { paths: mapPathToMethods };
-
-            const mapPathToMethodsYaml = yaml.dump(finalObj, { indent: 4, lineWidth: -1, noRefs: true });
-            const componentsYaml = yaml.dump(components, { indent: 4, lineWidth: -1, noRefs: true });
-
-            docs += mapPathToMethodsYaml;
-            docs += componentsYaml;
-
-            docs += `
-    securitySchemes:
-        AccessTokenAuth:
-            name: X-Access-Token
-            type: apiKey
-            in: header
-            description: |-
-                If authentication is enabled in the WildDuck configuration, you will need to supply an access token in the \`X-Access-Token\` header.
-
-                \`\`\`json
-                {
-                    "X-Access-Token": "59fc66a03e54454869460e45"
-                }
-                \`\`\`
-security:
-    - AccessTokenAuth: []
-`;
-
-            await fs.promises.writeFile(__dirname + '/openapidocs.yml', docs);
+            tools.generateAPiDocs(routes);
         })
     );
-
-    // ignore function and symbol types
-    const joiTypeToOpenApiTypeMap = {
-        any: 'object',
-        number: 'number',
-        link: 'string',
-        boolean: 'boolean',
-        date: 'string',
-        string: 'string',
-        binary: 'string'
-    };
-
-    function replaceWithRefs(reqBodyData) {
-        if (reqBodyData.type === 'array') {
-            const obj = reqBodyData.items[0];
-
-            replaceWithRefs(obj);
-        } else if (reqBodyData.type === 'object') {
-            if (reqBodyData.objectName) {
-                const objectName = reqBodyData.objectName;
-                Object.keys(reqBodyData).forEach(key => {
-                    if (key !== '$ref') {
-                        delete reqBodyData[key];
-                    }
-                });
-                reqBodyData.$ref = `#/components/schemas/${objectName}`;
-            } else {
-                for (const key in reqBodyData.properties) {
-                    replaceWithRefs(reqBodyData.properties[key]);
-                }
-            }
-        } else if (reqBodyData.type === 'alternatives') {
-            for (const obj in reqBodyData.oneOf) {
-                replaceWithRefs(obj);
-            }
-        }
-    }
-
-    function parseComponetsDecoupled(component, components) {
-        if (component.type === 'array') {
-            const obj = structuredClone(component.items[0]); // copy
-
-            if (obj.objectName) {
-                for (const key in obj.properties) {
-                    parseComponetsDecoupled(obj.properties[key], components);
-                }
-
-                // in case the Array itself is marked as a separate object >
-                const objectName = obj.objectName;
-                components[objectName] = obj;
-                delete components[objectName].objectName;
-                // ^
-            }
-        } else if (component.type === 'object') {
-            const obj = structuredClone(component); // copy
-            const objectName = obj.objectName;
-
-            for (const key in obj.properties) {
-                parseComponetsDecoupled(obj.properties[key], components);
-            }
-
-            if (objectName) {
-                components[objectName] = obj;
-                delete components[objectName].objectName;
-            }
-        } else if (component.oneOf) {
-            // Joi object is of 'alternatives' types
-            for (const obj in component.oneOf) {
-                parseComponetsDecoupled({ ...obj }, components);
-            }
-        }
-    }
-
-    /**
-     * Parse Joi Objects
-     */
-    function parseJoiObject(path, joiObject, requestBodyProperties) {
-        if (joiObject.type === 'object') {
-            const fieldsMap = joiObject._ids._byKey;
-
-            const data = {
-                type: joiObject.type,
-                description: joiObject._flags.description || '',
-                properties: {},
-                required: []
-            };
-
-            if (joiObject._flags.objectName) {
-                data.objectName = joiObject._flags.objectName;
-            }
-
-            if (path) {
-                requestBodyProperties[path] = data;
-            } else {
-                requestBodyProperties.push(data);
-            }
-
-            for (const [key, value] of fieldsMap) {
-                if (value.schema._flags.presence === 'required') {
-                    data.required.push(key);
-                }
-                parseJoiObject(key, value.schema, data.properties);
-            }
-        } else if (joiObject.type === 'alternatives') {
-            const matches = joiObject.$_terms.matches;
-
-            const data = {
-                oneOf: [],
-                description: joiObject._flags.description || ''
-            };
-
-            if (path) {
-                requestBodyProperties[path] = data;
-            } else {
-                requestBodyProperties.push(data);
-            }
-
-            for (const alternative of matches) {
-                parseJoiObject(null, alternative.schema, data.oneOf);
-            }
-        } else if (joiObject.type === 'array') {
-            const elems = joiObject?.$_terms.items;
-
-            const data = {
-                type: 'array',
-                items: [],
-                description: joiObject._flags.description || ''
-            };
-
-            if (path) {
-                requestBodyProperties[path] = data;
-            } else {
-                requestBodyProperties.push(data);
-            }
-            parseJoiObject(null, elems[0], data.items);
-        } else {
-            const openApiType = joiTypeToOpenApiTypeMap[joiObject.type]; // even if type is object here then ignore and do not go recursive
-            const isRequired = joiObject._flags.presence === 'required';
-            const description = joiObject._flags.description || '';
-            let format = undefined;
-
-            if (!openApiType) {
-                throw new Error('Unsupported type! Check API endpoint!');
-            }
-
-            if (joiObject.type !== openApiType) {
-                // type has changed, so probably string, acquire format
-                format = joiObject.type;
-            }
-
-            const data = { type: openApiType, description, required: isRequired };
-            if (format) {
-                data.format = format;
-            }
-            if (path) {
-                requestBodyProperties[path] = data;
-            } else {
-                requestBodyProperties.push(data);
-            }
-        }
-    }
 
     server.on('error', err => {
         if (!started) {
